requirements.txt
openai==1.20.0pip
python-dotenv